python-app/app/ai_agent/requirements.txt
chromadb
pypdf
python-docx
PyPDF2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
dataclasses-json==0.6.7
distro==1.9.0
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.1
flask_cors
frozenlist==1.6.0
fsspec==2025.5.0
greenlet==3.2.2
h11==0.16.0
huggingface-hub==0.31.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.60
langchain-huggingface==0.2.0
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-checkpoint-sqlite==2.0.10
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.69
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.4
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.6
openai==1.79.0
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pillow==11.2.1
propcache==0.3.1
pydantic==2.11.4
pydantic-settings==2.9.1
docx2txt
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlite-vec==0.1.6
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.2.2
tqdm==4.67.1
transformers==4.52.2
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0

# OCR dependencies
pytesseract
pdf2image
pillow
